user management functionalities added

+++ src/app/app-settings/components/role-management/role-management.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from "@angular/core";
-import { MenuItem, MessageService } from "primeng/api";
+import {ConfirmationService, MenuItem, MessageService} from "primeng/api";
 import { Table } from "primeng/table";
 import { RoleSettingsService } from "../../services/role-settings.service"
 import  {AuthenticationService} from "../../../auth/services/authentication.service";
@@ -14,6 +14,7 @@ import {CheckLoginService} from "../../../shared/shared-services/check-login.ser
   selector: "app-role-management",
   templateUrl: "./role-management.component.html",
   styleUrl: "./role-management.component.scss",
+  providers: [MessageService, ConfirmationService]
 })
 export class RoleManagementComponent implements OnInit{
 
@@ -22,8 +23,11 @@ export class RoleManagementComponent implements OnInit{
     { label: "Role Management" },
   ];
 
-  roles: {group_name:string, number_of_users:number}[] = []
+  roles!: {group_name:string, number_of_users:number}[];
   selectedRoles!: { group_name: string, number_of_users: number };
+  viewRolePopUpVisible: boolean = false;
+  userGroupData:any;
+
 
   actions!: MenuItem[];
 
@@ -33,7 +37,9 @@ export class RoleManagementComponent implements OnInit{
     private roleRefreshService: RoleRefreshService,
     private messageService: MessageService,
     private router: Router,
-    private checkLoginService: CheckLoginService
+    private checkLoginService: CheckLoginService,
+    private confirmationService: ConfirmationService
+
   ) {}
 
 
@@ -55,6 +61,13 @@ export class RoleManagementComponent implements OnInit{
         icon: "pi pi-eye",
         command: () => {
           console.log("Viewing role");
+          this.getRoleDetails()
+
+          if(this.selectedRoles){
+            this.viewRole();
+          }else {
+            this.messageService.add({severity: 'error', summary: 'Error', detail: 'No role selected'});
+          }
         }
       },
       { label: "Update",
@@ -68,11 +81,21 @@ export class RoleManagementComponent implements OnInit{
         icon: "pi pi-trash",
         command: () => {
           console.log("Deleting role");
-          this.deleteRole();
+          if(this.selectedRoles) {
+            this.deleteRole();
+          }else{
+            this.messageService.add({severity: 'error', summary: 'Error', detail: 'No role selected'});
+          }
         }
       },
     ];
 
+  }
+
+  viewRole() {
+    // Here you can fetch more details about the role if needed.
+    // For now, we'll just make the role view popup visible.
+    this.viewRolePopUpVisible = true;
   }
 
   refreshRoles() {
@@ -84,34 +107,59 @@ export class RoleManagementComponent implements OnInit{
     });
   }
 
-deleteRole() {
-  this.authService.getIdToken().pipe(
-    catchError(error => {
-      if (error === 'Session expired. Please sign in again.') {
-        this.router.navigate(['/auth/login']);
-      } else {
-        this.messageService.add({severity:'error', summary:'Error', detail:'Failed to get token'});
+  deleteRole() {
+    this.confirmationService.confirm({
+      header: 'Are you sure?',
+      message: 'Please confirm to proceed.',
+      accept: () => {
+        this.authService.getIdToken().pipe(
+          catchError(error => {
+            if (error === 'Session expired. Please sign in again.') {
+              this.router.navigate(['/auth/login']);
+            } else {
+              this.messageService.add({severity:'error', summary:'Error', detail:'Failed to get token'});
+            }
+            return of(null);
+          })
+        ).subscribe(token => {
+          if (token) {
+            console.log(this.selectedRoles.group_name)
+            this.roleService.deleteUserRole(token, this.selectedRoles.group_name).subscribe(
+              (data: any) => {
+                console.log(data);
+                this.refreshRoles();
+                this.messageService.add({severity:'success', summary: 'Success', detail: 'Role Deleted Successfully'});
+              },
+              (error: any) => {
+                // Handle the error case
+                console.log(error);
+                this.messageService.add({severity:'error', summary: 'Error', detail: 'Failed to delete role'});
+              }
+            );
+          }
+        });
+      },
+      reject: () => {
+        this.messageService.add({ severity: 'error', summary: 'Rejected', detail: 'You have rejected', life: 3000 });
       }
-      return of(null);
-    })
-  ).subscribe(token => {
-    if (token) {
-      console.log(this.selectedRoles.group_name)
-      this.roleService.deleteUserRole(token, this.selectedRoles.group_name).subscribe(
+    });
+  }
+
+  getRoleDetails() {
+    this.authService.getIdToken().subscribe((token: any) => {
+      this.roleService.getRoleDetails(token, this.selectedRoles.group_name).subscribe(
         (data: any) => {
           console.log(data);
-          this.refreshRoles();
-          this.messageService.add({severity:'success', summary: 'Success', detail: 'Role Deleted Successfully'});
+          this.userGroupData = data;
         },
         (error: any) => {
-          // Handle the error case
           console.log(error);
-          this.messageService.add({severity:'error', summary: 'Error', detail: 'Failed to delete role'});
+          this.messageService.add({severity: 'error', summary: 'Error', detail: 'Failed to get role details'});
         }
       );
-    }
-  });
-}
+    });
+
+  }
 
 
   // roles = [
